bg(refactor):changed utils directory
